dto 수정 - AuthorBrief는 dto 패키지로, 기타 DTO 이름 수정
