+++ src/main/kotlin/com/waffletoy/team1server/post/controller/PostController.kt
@@ -1,13 +1,16 @@
 package com.waffletoy.team1server.post.controller
 
+import com.amazonaws.HttpMethod
 import com.waffletoy.team1server.post.dto.Post
 import com.waffletoy.team1server.post.dto.PostBrief
 import com.waffletoy.team1server.post.service.PostService
+import com.waffletoy.team1server.post.service.S3Service
 import com.waffletoy.team1server.user.AuthUser
 import com.waffletoy.team1server.user.dtos.User
 import io.swagger.v3.oas.annotations.Parameter
 import jakarta.validation.constraints.Max
 import jakarta.validation.constraints.Min
+import org.springframework.beans.factory.annotation.Value
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.validation.annotation.Validated
@@ -18,6 +21,8 @@ import org.springframework.web.bind.annotation.*
 @Validated
 class PostController(
     private val postService: PostService,
+    private val s3Service: S3Service,
+    @Value("\${amazon.aws.bucket}") private val bucketName: String,
 ) {
     // 채용 공고 상세 페이지 불러오기
     @GetMapping("/{post_id}")
@@ -37,7 +42,7 @@ class PostController(
         @RequestParam(required = false) @Min(0) @Max(2) status: Int?,
         @RequestParam(required = false) series: List<String>?,
         @RequestParam(required = false) @Min(0) page: Int?,
-    ): ResponseEntity<PostWithPageDTO> {
+    ): ResponseEntity<PostWithPage> {
         val posts = postService.getPosts(roles, investmentMax, investmentMin, status, series, page ?: 0)
 
         // 총 페이지
@@ -45,7 +50,7 @@ class PostController(
 
         // PostBrief로 매핑하여 반환
         return ResponseEntity.ok(
-            PostWithPageDTO(
+            PostWithPage(
                 posts = posts.content.map { PostBrief.fromPost(Post.fromEntity(it)) },
                 paginator = Paginator(totalPages),
             ),
@@ -77,7 +82,7 @@ class PostController(
     fun getBookMarks(
         @Parameter(hidden = true) @AuthUser user: User,
         @RequestParam(required = false) @Min(0) page: Int?,
-    ): ResponseEntity<PostWithPageDTO> {
+    ): ResponseEntity<PostWithPage> {
         val posts = postService.getBookmarks(user.id, page ?: 0)
 
         // 총 페이지
@@ -85,7 +90,7 @@ class PostController(
 
         // PostBrief로 매핑하여 반환
         return ResponseEntity.ok(
-            PostWithPageDTO(
+            PostWithPage(
                 posts = posts.content.map { PostBrief.fromPost(Post.fromEntity(it)) },
                 paginator = Paginator(totalPages),
             ),
@@ -111,21 +116,44 @@ class PostController(
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build()
         }
     }
-}
 
-data class AuthorBriefDTO(
-    val id: String,
-    val name: String,
-    val profileImageLink: String?,
-)
+    @PostMapping("/upload/presigned")
+    fun generateUploadPresignedUrl(
+        @RequestBody preSignedUploadReq: PreSignedUploadReq,
+    ): ResponseEntity<PresignedURL> {
+        val presignedUrl = s3Service.generateUploadPreSignUrl(preSignedUploadReq, bucketName, HttpMethod.PUT)
+        return ResponseEntity.ok(PresignedURL(presignedUrl))
+    }
+
+    @PostMapping("/download/presigned")
+    fun generateDownloadPresignedUrl(
+        @RequestBody preSignedDownloadReq: PreSignedDownloadReq,
+    ): ResponseEntity<PresignedURL> {
+        val presignedUrl = s3Service.generateDownloadPreSignUrl(preSignedDownloadReq, bucketName)
+        return ResponseEntity.ok(PresignedURL(presignedUrl))
+    }
+}
 
 data class Paginator(
     val lastPage: Int,
 )
 
-data class PostWithPageDTO(
+data class PostWithPage(
     val posts: List<PostBrief>,
     val paginator: Paginator,
+)
+
+data class PreSignedUploadReq(
+    val fileName: String,
+    val fileType: String,
+)
+
+data class PreSignedDownloadReq(
+    val fileName: String,
+)
+
+data class PresignedURL(
+    val presignedUrl: String,
 )
 
 data class PasswordRequest(val pw: String)
