+++ src/main/kotlin/com/waffletoy/team1server/post/dto/AuthorBrief.kt
@@ -1,0 +1,7 @@
+package com.waffletoy.team1server.post.dto
+
+data class AuthorBrief(
+    val id: String,
+    val name: String,
+    val profileImageLink: String?,
+)
