+++ src/main/kotlin/com/waffletoy/team1server/post/dto/PostBrief.kt
@@ -2,12 +2,11 @@ package com.waffletoy.team1server.post.dto
 
 import com.waffletoy.team1server.post.Category
 import com.waffletoy.team1server.post.Series
-import com.waffletoy.team1server.post.controller.AuthorBriefDTO
 import java.time.LocalDateTime
 
 data class PostBrief(
     val id: String,
-    val author: AuthorBriefDTO,
+    val author: AuthorBrief,
     val companyName: String,
     val slogan: String,
     val email: String,
